Move classes in folders
