+++ RateLimiter/RuleTemplates/RegionBasedRuleTemplate.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using RateLimiter.Rules;
+using RateLimiter.Rules.Constructors;
 using RateLimiter.RuleTemplates.Params;
 
 namespace RateLimiter.RuleTemplates
