+++ RateLimiter/Rules/Constructors/RequestByTimeSpanRuleConstructor.cs
@@ -1,8 +1,8 @@
-﻿using RateLimiter.Repositories;
-using RateLimiter.Rules;
+﻿using RateLimiter.Rules;
+using RateLimiter.RuleTemplates;
 using RateLimiter.RuleTemplates.Params;
 
-namespace RateLimiter.RuleTemplates
+namespace RateLimiter.Rules.Constructors
 {
     internal class RequestByTimeSpanRuleConstructor : IRuleConstructor
     {
@@ -11,7 +11,7 @@ namespace RateLimiter.RuleTemplates
             if (templateParams is not RequestByTimeSpanRuleTemplateParams validParams)
             {
                 throw new InvalidTemplateParamsException();
-                
+
             }
             return new RequestByTimeSpanRule(validParams.RequestLimit, validParams.TimeSpan);
         }
